backend/requirements.txt
Django>=5.2.6,<6.0
djangorestframework>=3.14.0,<4.0
celery>=5.3.0,<6.0
redis>=4.5.0,<5.0
django-celery-beat>=2.5.0,<3.0
django-celery-results>=2.5.0,<3.0
psycopg2-binary>=2.9.0,<3.0
pytest>=7.4.0,<8.0
pytest-django>=4.5.0,<5.0
freezegun>=1.2.0,<2.0
playwright>=1.35.0,<2.0